refactor(app): clean app.js
Refactored language to use useState hook instead of a let variable.
Moved the language detection logic inside useEffect to be executed only on client-side.
Renamed appProps to pageProps to match Next.js documentation.
Removed the import of Head component as it's not being used.
Removed the import of HeadContent component as it's being included directly in MyApp component