refactor(document): modern way
getInitialProps is now defined as a static method on the function component, rather than a static method on the class component. The sheet is now passed to getInitialProps via the props parameter, rather than being declared in the function body. Finally, getInitialProps now returns styleTags as a separate prop, which is used in the Head component to output the styles.